Addressing feedback from @BlakeWilliams.
