+++ lib/slack.ex
@@ -77,6 +77,8 @@ defmodule Slack do
       @behaviour :websocket_client_handler
       import Slack
       import Slack.Handlers
+      import Slack.Lookups
+      import Slack.Sends
 
       def start_link(token, initial_state, client \\ :websocket_client) do
         case Slack.Rtm.start(token) do
@@ -163,134 +165,5 @@ defmodule Slack do
 
       defoverridable [handle_connect: 2, handle_message: 3, handle_close: 3, handle_info: 3]
     end
-  end
-
-  @doc ~S"""
-  Turns a string like `"@USER_NAME"` into the ID that Slack understands (`"U…"`).
-  """
-  def lookup_user_id("@" <> user_name, slack) do
-    slack.users
-    |> Map.values
-    |> Enum.find(%{ }, fn user -> user.name == user_name end)
-    |> Map.get(:id)
-  end
-
-  @doc ~S"""
-  Turns a string like `"@USER_NAME"` or a user ID (`"U…"`) into the ID for the
-  direct message channel of that user (`"D…"`).  `nil` is returned if a direct
-  message channel has not yet been opened.
-  """
-  def lookup_direct_message_id(user = "@" <> _user_name, slack) do
-    user
-    |> lookup_user_id(slack)
-    |> lookup_direct_message_id(slack)
-  end
-  def lookup_direct_message_id(user_id, slack) do
-    slack.ims
-    |> Map.values
-    |> Enum.find(%{ }, fn direct_message -> direct_message.user == user_id end)
-    |> Map.get(:id)
-  end
-
-  @doc ~S"""
-  Turns a string like `"@CHANNEL_NAME"` into the ID that Slack understands
-  (`"C…"`).
-  """
-  def lookup_channel_id("#" <> channel_name, slack) do
-    slack.channels
-    |> Map.values
-    |> Enum.find(fn channel -> channel.name == channel_name end)
-    |> Map.get(:id)
-  end
-
-  @doc ~S"""
-  Turns a Slack user ID (`"U…"`) or direct message ID (`"D…"`) into a string in
-  the format "@USER_NAME".
-  """
-  def lookup_user_name(direct_message_id = "D" <> _id, slack) do
-    lookup_user_name(slack.ims[direct_message_id].user, slack)
-  end
-  def lookup_user_name(user_id = "U" <> _id, slack) do
-    "@" <> slack.users[user_id].name
-  end
-
-  @doc ~S"""
-  Turns a Slack channel ID (`"C…"`) into a string in the format "#CHANNEL_NAME".
-  """
-  def lookup_channel_name(channel_id = "C" <> _id, slack) do
-    "#" <> slack.channels[channel_id].name
-  end
-
-  @doc """
-  Sends `text` to `channel` for the given `slack` connection.  `channel` can be
-  a string in the format of `"#CHANNEL_NAME"`, `"@USER_NAME"`, or any ID that
-  Slack understands.
-  """
-  def send_message(text, channel = "#" <> channel_name, slack) do
-    channel_id = lookup_channel_id(channel, slack)
-    if channel_id do
-      send_message(text, channel_id, slack)
-    else
-      raise ArgumentError, "channel ##{channel_name} not found"
-    end
-  end
-  def send_message(text, user = "@" <> user_name, slack) do
-    direct_message_id = lookup_direct_message_id(user, slack)
-    if direct_message_id do
-      send_message(text, direct_message_id, slack)
-    else
-      im_open = HTTPoison.post(
-        "https://slack.com/api/im.open",
-        {:form, [token: slack.token, user: lookup_user_id(user, slack)]}
-      )
-      case im_open do
-        {:ok, response} ->
-          case JSX.decode!(response.body, [{:labels, :atom}]) do
-            %{ok: true, channel: %{id: id}} -> send_message(text, id, slack)
-            _ -> :delivery_failed
-          end
-        {:error, reason} -> :delivery_failed
-      end
-    end
-  end
-  def send_message(text, channel, slack) do
-    %{
-      type: "message",
-      text: text,
-      channel: channel
-    }
-      |> JSX.encode!
-      |> send_raw(slack)
-  end
-
-  @doc """
-  Notifies Slack that the current user is typing in `channel`.
-  """
-  def indicate_typing(channel, slack) do
-    %{
-      type: "typing",
-      channel: channel
-    }
-      |> JSX.encode!
-      |> send_raw(slack)
-  end
-
-  @doc """
-  Notifies slack that the current `slack` user is typing in `channel`.
-  """
-  def send_ping(data \\ [], slack) do
-    %{
-      type: "ping"
-    }
-      |> Dict.merge(data)
-      |> JSX.encode!
-      |> send_raw(slack)
-  end
-
-  @doc """
-  Sends raw JSON to a given socket.
-  """
-  def send_raw(json, %{socket: socket, client: client}) do
-    client.send({:text, json}, socket)
   end
 end
